+++ NEWS.md
@@ -3,7 +3,10 @@
 - implemented a type-based distinction of the two modes of `grid_x` and `grid_y`: Integer grid length mode has to be triggered explicitly now by giving an integer with a trailing L, e.g. `10L` (this is a breaking change!)
 - added two new geoms `geom_labelgrid()` and `geom_segmentgrid()`
 - defined a custom ggplot2 stat `stat_grid_arrange()` that does the hard work of the grid arrangement for the geoms
-- implemented an interface for `stat_grid_arrange()` and therefore the geoms, which allows to set grid parameters per axis, or directly provide the grid, and set polygons to which the grid is filtered
+- implemented an interface for `stat_grid_arrange()` and therefore the geoms, which allows to
+    - set grid parameters per axis or directly provide the grid
+    - set polygons to which the grid is filtered
+    - rescale the axes to improve the arrangement if x and y are on very different numeric scales
 - added a function `filter_grid_in_polygons()` to filter a grid to those points falling inside polygonal regions supplied as sf geometries (sf is a suggested dependency now)
 - remodelled the calling of the futhark code to support a second module underpinning `filter_grid_in_polygons()`
 - switched to the latest futhark version 0.26.1, which required minor adjustments of the Rcpp bridge code